feat: 캠프리스트 캐러셀 반응형 여백 유지 추가, 홈 페이지 메타데이터 추가
